+++ app/src/main/res/layout/fragment_login.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple"
-    tools:context=".LoginFragment">
+    tools:context=".login.LoginFragment">
 
     <ImageView
         android:id="@+id/leftImage"
