+++ app/src/main/res/layout/fragment_register.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/purple"
-    tools:context=".RegisterFragment">
+    tools:context=".register.RegisterFragment">
 
     <ImageView
         android:id="@+id/leftImage"
