+++ app/src/main/res/layout/fragment_show_details.xml
@@ -11,7 +11,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".ShowDetailsFragment">
+        tools:context=".show_details.ShowDetailsFragment">
 
         <ImageView
             android:id="@+id/detailsBackButton"
