+++ app/src/main/res/layout/fragment_shows.xml
@@ -5,13 +5,13 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".ShowsFragment">
+    tools:context=".shows.ShowsFragment">
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:background="@color/white"
-        tools:context=".ShowsFragment">
+        tools:context=".shows.ShowsFragment">
 
         <ImageView
             android:id="@+id/showsUser"
