+++ app/src/main/res/navigation/main.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/loginFragment"
-        android:name="com.example.shows_saratedd.LoginFragment"
+        android:name="com.example.shows_saratedd.login.LoginFragment"
         tools:layout="@layout/fragment_login">
         <action
             android:id="@+id/toShowsFragment"
@@ -19,7 +19,7 @@
 
     <fragment
         android:id="@+id/registerFragment"
-        android:name="com.example.shows_saratedd.RegisterFragment"
+        android:name="com.example.shows_saratedd.register.RegisterFragment"
         tools:layout="@layout/fragment_register">
         <action
             android:id="@+id/toLoginFragment"
@@ -28,7 +28,7 @@
 
     <fragment
         android:id="@+id/showsFragment"
-        android:name="com.example.shows_saratedd.ShowsFragment"
+        android:name="com.example.shows_saratedd.shows.ShowsFragment"
         tools:layout="@layout/fragment_shows">
         <argument
             android:name="email"
@@ -43,7 +43,7 @@
 
     <fragment
         android:id="@+id/showDetailsFragment"
-        android:name="com.example.shows_saratedd.ShowDetailsFragment"
+        android:name="com.example.shows_saratedd.show_details.ShowDetailsFragment"
         tools:layout="@layout/fragment_show_details">
         <argument
             android:name="showName"
